+++ packages/sitecore-jss-react/src/components/Link.tsx
@@ -25,7 +25,7 @@ export interface LinkField {
   editableLastPart?: string;
 }
 
-export type LinkProps = EditableFieldProps &
+export type LinkProps = EditableFieldProps<LinkProps> &
   React.AnchorHTMLAttributes<HTMLAnchorElement> &
   RefAttributes<HTMLAnchorElement> & {
     /** The link field data. */