+++ packages/sitecore-jss-react/src/enhancers/withEmptyFieldEditingComponent.test.tsx
@@ -56,6 +56,7 @@ describe('withEmptyFieldEditingComponent', () => {
 
     it('Should render provided default empty value component component if field value is not provided', () => {
       const props = {
+        tag: 'h1',
         field: {
           value: '',
           metadata: testMetadata,
@@ -67,9 +68,10 @@ describe('withEmptyFieldEditingComponent', () => {
       });
 
       const rendered = render(<WrappedComponent {...props} />);
-      const expected = render(<DefaultEmptyFieldEditingComponentText />);
+      const expected = render(<DefaultEmptyFieldEditingComponentText tag="h1" />);
 
       expect(rendered.container.innerHTML).to.equal(expected.container.innerHTML);
+      expect(rendered.container.innerHTML).to.equal('<h1 tag="h1">[No text in field]</h1>');
     });
 
     it('Should render custom empty value component if provided via props if field value is not provided', () => {